requirements.txt
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
xgboost==2.1.1
matplotlib==3.8.2
apache-airflow==2.10.2